Check for ES6Promise and ES6Proxy in original trace when replaying
Summary:
If the RuntimeConfig fields "ES6Promise" or "ES6Proxy" exist in the
original trace, replay with those fields set as well.
This avoids accidentally running a trace with Proxy turned on when it was
originally taken with Proxy turned off.
Recording is already implemented in `SynthTrace.cpp`, this change handles replay

In practice this didn't cause any issues, but in general features like Proxy might
have changed the sequence of API actions, and should be recorded and replayed
properly.

Reviewed By: kodafb

Differential Revision: D26349818

fbshipit-source-id: 540733f9ce0d7c4db82c08f7c380d9c231039ab3