Use CollectMessagesRAII when backtracking.
Summary:
When performing certain kinds of backtracking, such as required in
```
x ? (1) : <foo>#{bar}</foo>
```
upon encountering the `:`, the lexer must attempt to parse type parameters
in `<foo>`, and then check the next token to see if a function type
is actually what is being parsed.

If it's actually JSX being parsed, then the lexing rules are _different_
than regular JS, and any errors we encounter when lexing the next token
as JS must be suppressed. However, if it actually did end up parsing as
JS, we must register and print the lexer errors.

To do this, introduce a `CollectMessagesRAII` RAII class which
registers itself as an "external message buffer" in `SourceErrorManager`.
The user of `CollectMessagesRAII` can alter the destruction behavior
by setting a flag to determine whether it should display any lexer
messages or simply discard them.

Reviewed By: tmikov

Differential Revision: D24933670

fbshipit-source-id: cf7d060059caf4f1b4fc5b857b5103280a747869