Report position of parse error on error object
Summary:
When ESLint's parser throws a parse error, ESLint expects that the parse error object will have `lineNumber` and `column` properties set to indicate the source location of the parse error. Let's make sure to set these properties on parse errors in `hermes-eslint` so that parse errors are properly reported.

First we must thread the error line and column from the WASM parser's diagnostic handler through to JS, where the line and column can be fetched via the new `hermesParseResult_getErrorLine` and `hermesParseResult_getErrorColumn` functions. The WASM parser then can save them under a `loc` key on the error object (this is the same format used by Babel). In `hermes-eslint` we then unwrap this `loc` key and set the `lineNumber` and `column` properties of the error object. We don't want to set these properties for errors not produced by the `hermes-eslint` package since `lineNumber` is a property of SyntaxErrors that is already used.

Reviewed By: pieterv

Differential Revision: D26350223

fbshipit-source-id: 4cd797ef412984dfa86c15a345e88d9cb7a53c73