Enforce error handling in Regex Executor
Summary:
Mark functions that push onto the backtrack stack as NODISCARD to
enforce that we actually check the result.

Convert `backtrack` and `matchWidth1Loop` to return `OptValue`
because they are already using their return value for something else.

Add an assert to catch uncaught errors on the next iteration of the
loop.

Reviewed By: avp

Differential Revision: D26379686

fbshipit-source-id: 5b834ce70b3c46c8d680aadc9c4c3ffa3c791392