Enable compressed pointers for RuntimeGC
Summary:
Runtime GC supports compressed pointers, in order to get a good heap
size comparison, we should enable it.

This should also fix the production crash with concurrent marking by
forcing pointers to be assigned to an intermediate variable before being
marked. However, we still need to fix the case for when compressed
pointers are disabled.

Reviewed By: dulinriley

Differential Revision: D26417155

fbshipit-source-id: de9d5dae44a00d3f3ba8aefaed14799c7dc9f21c