Implement an executor to use in Hades
Summary:
Instead of creating a new Hades thread for each collection, we can
keep the same thread around for the entire life of the runtime. This
saves us the thread creation cost at each full collection.

The executor can also be used to make simplifications to how the
mutator synchronises with the background thread.

Finally, since this executor API is fairly simple, we could expose
a config option to users to supply their own executor, so we can
reuse existing threads.

As a nice side-effect, it's also useful for profiling Hades, since
all work happens on a single thread.

Reviewed By: dulinriley

Differential Revision: D26299451

fbshipit-source-id: 6b2fe24f7e70cfbf4a9f4b1f67725dacd56d42b2