Use executor to replace STW cond var
Summary:
If we want to use an externally provided executor, we should never
block it by waiting on a condition variable. Instead, restructure the
collection loop to not need to block on the STW cond var.

The background thread should exit as soon as we hit the
`CompleteMarking` phase, and then the mutator will schedule the
sweep phase after the STW pause is completed

Reviewed By: dulinriley

Differential Revision: D26299515

fbshipit-source-id: ad029ac9e8c5cade048ea48d9d676f5c2568d2d5