Add GCBase::getKindAsStr to formalize GC name fetching
Summary:
Formalize the "name" of the GC via a new method `getKindAsStr`.
This will ensure all 3 of:
* Crash info
* Analytics logging
* `HermesInternal.getRuntimeProperties()`

agree on the correct GC name.
Particularly for Hades, where experiments can be enabled, we want to make
sure the crash data includes which experiments are turned on.
Similarly, for testing, it's helpful to expose the Hades experiments in the
RN Hub

Reviewed By: neildhar

Differential Revision: D26564054

fbshipit-source-id: a4127e6fd8401e8c07f31a117091b04fcf6c5d99