Singleton all CreatePropNameIDRecords in normalization
Summary:
The `CreatePropNameIDRecord` creates a SymbolID for a
native string. This SymbolID might be the same for the same
PropNameID, unless a GC has occurred, in which case it might
be a different SymbolID.

The problem with that is that it encodes some knowledge of GC
scheduling into the trace, such that a replay might not have the
same characteristics. When comparing a retrace and the original,
we need to fix these records.

To do so, assume that a CreatePropNameIDRecord can always
choose to use the same ID. Map all CreatePropNameIDRecords
with the same string contents to the same SymbolID.

This means keeping the symbol alive
longer than it may have before normalization. For this reason a
normalized trace shouldn't be checked in, as it's slightly different
and possibly retaining more memory than the original. However,
it's still the best way to compare an original trace and a retrace to
ensure everything was replayed correctly.

Reviewed By: avp

Differential Revision: D26499314

fbshipit-source-id: 3aafa486fa00c1659586fbf882c00678cac127aa