Fix VM serialisation when compressed pointers are off
Summary:
Add the ability to convert back and forth between the storage type of
GCPointer and a raw representation. This fixes a bug where
serialisation is broken if compressed pointers are turned off.

The ability to convert back and forth is also used in HV32, so more
use cases for both of these functions will be added soon.

Reviewed By: dulinriley

Differential Revision: D26653487

fbshipit-source-id: b8b6587f6e92ee408219f8585629a5b884d0d357