Fix GenGCHeapSegment crash manager extents to be valid JSON
Summary:
Make sure the string output via crash manager is encoded in a way where it
can be embedded as a string inside JSON correctly.
This means adding quotes around `lo` and `hi`, and escaping the quotes.

Also moved `addExtentToString` to `GenGCHeapSegment` since it was only
used by GenGC anyway.

Reviewed By: neildhar

Differential Revision: D26677848

fbshipit-source-id: 085ca2ac31f03b93419d5a2a55f441ed43edf2d6