Simplify CMake build rules for synth trace (#462)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/462

We currently fail to build `hermesAPI` with CMake when synth traces are
enabled because we do not link `hermesAPI` against
`hermesInstrumentation`. Fix this by matching our CMake build more
closely with our Buck build, where we have a separate
`TraceInterpreter` library, that is then linked against by anything
that needs to use it.

We can rely on the ifdefs inside the cpp files to just omit the
source code when tracing is disabled.

Reviewed By: dulinriley

Differential Revision: D26760317

fbshipit-source-id: c9e55c22b642b76aafe8a71b7401607c5677ca3f