Silence warnings when compiling bytecode using WASM.
Summary:
When using `hermesc.js` to compile to bytecode it was displaying some unwanted warnings about `pthread_sigmask`. For example:
```
[mns@devvm111.cln0 ~/fbsource] node buck-out/gen/ce9b6f2e/xplat/hermes/tools/hermesc/hermesc.js -emit-binary -out ~/foo.hbc ~/foo.js
pthread_sigmask() is not supported: this is a no-op.
pthread_sigmask() is not supported: this is a no-op.
```
With this diff, `LLVM_ENABLE_THREADS` is set to 0 in the WASM config. This means prevents it from trying to use the function `pthread_sigmask`:

https://www.internalfb.com/intern/diffusion/FBS/browsefile/master/xplat/hermes/external/llvh/lib/Support/Unix/Process.inc?commit=504758d2ae39ed63e7ffdf81a390c796583509ce&lines=236-242%2C251-256

This seems like the most non-intrusive way to silence the warning. Let me know if there is concern this might affect performance and we can find another way to handle this.

Reviewed By: dulinriley

Differential Revision: D26690011

fbshipit-source-id: d092f4f642316c8e338f1ddcd386a769146d018e