Add option to disable source hash check
Summary:
Add an option that removes the requirement that the input bytecode was compiled from the same source used to record the trace. There must only be one input bytecode file in this case. If its observable behavior deviates from the trace, the results are undefined.

This option is intended to be used when testing/debugging changes to Hermes that involve changes to source transformations.

#utd-hermes-ignore-windows

Reviewed By: dulinriley

Differential Revision: D26793935

fbshipit-source-id: b1747ec4c1db2bd5d91f472db7d4ecd40d21bb8c