Fix CMake build with exceptions enabled
Summary:
We should build with exceptions when `HERMESVM_EXCEPTION_ON_OOM` is
set. Building with this flag is currently broken.

Note that after this change, `compileJS` will still be built without
exceptions, because we explicitly disable exceptions for it. We
could save and restore default value for `compileJS` but that
probably isn't very useful, given that it can't actually throw.

While testing this, I also found that one of our tests has gone out
of date, so that is updated here as well.

Reviewed By: tmikov

Differential Revision: D26802764

fbshipit-source-id: a8ce45faf5abc8b586c5f7b580db99173bb65cde