Delete HERMESVM_API_TRACE (#466)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/466

For buck, this change makes no difference, since the flag is set to
true by default.

For CMake, this means that we will be able to build the synth tool or
create synth traces without needing to specifically set this build
flag. More importantly, it ensures that the CMake build doesn't break
because it will be continuously built through CI.

There should be no effect on the binary size of hermesapi or
libhermes because the synth trace targets are not linked into them.

Reviewed By: dulinriley

Differential Revision: D26771883

fbshipit-source-id: 1332e39d6cc4ed20b378ab9da41c4be8c9ee90bd