Fix pause time tracking in Hades
Summary:
In order to correctly estimate the max pause time, we need to
distinguish collection time spent on the mutator from time spent in
the background thread. This diff makes the following changes:
1. Ignore the time for hades OG collections when calculating the max
pause time and total time spent in the GC. This ensures that we can
then also correctly calculate the mutator time by subtracting the
total GC time from the total overall time.
2. Record the time spent in `waitForCollectionToFinish` so that we
can count it towards the mutator time.

Reviewed By: dulinriley

Differential Revision: D26853388

fbshipit-source-id: c547844bbceb5248e6e01f26e7a84abd60bed4fc