Remove JSArray::shadowLength_
Summary:
This change removes the `shadowLength_` member of JSArray, and
instead uses the internal `length` property for all accesses.

Long ago, properties were all stored out of line in an ArrayStorage, and
keeping the length shadowed inside JSArray was a faster way to access it.
But now JSObject has some inline property space, one of which stores the
length. This length is encoded as a HermesValue, as a number. This means
decoding it is free (numbers have no mask or shift applied). Added
`JSObject::directSlotRef` for properties known to be in the direct slots,
to avoid a branch when none is required.

So, there's no need to store an extra `uint32_t` for this field. Due to the way
direct slots work and overlapping JSObject direct properties, there's no change
in memory from this. This is part of a larger plan to inline
a few elements into a JSArray, and shrinking the overlapping storage allows
fitting more elements inline.

Reviewed By: kodafb

Differential Revision: D26261734

fbshipit-source-id: f6b3ba057b226512e4b3cdf61b72abc64f2351f5