Add Flow types to scope manager Definitions and Variables
Summary:
Let's add some types to the scope manager in `hermes-eslint`. This diff focuses on adding Flow types to the `definitions.js` and `variables.js` files.

We can hard code a real type for `Identifier` nodes since they are commonly passed around in the scope manager and we would like to distinguish them. However we do not yet have type definitions for all other AST nodes so we leave them as the generic `Node` type for now (which is `any` under the hood).

Reviewed By: gkz

Differential Revision: D26997910

fbshipit-source-id: d4e029e27ac2e9ebee7ef86ceac409c3ba5f7316