Add Flow to parser and eslint packages
Summary:
Let's start adding Flow types to the `hermes-parser` and `hermes-eslint` source code. Flow can be run for both these packages at once with the command `yarn flow`, and the babel configuration has been updated to strip Flow types when compiling.

This diff then enables Flow in the root `index.js` files for `hermes-parser` and `hermes-eslint`, and adds Flow types to these files (more files will be flowified in the future, but starting here). Since the ESLint visitor keys are generated during the build step, we must add a type definition file for them so that Flow is aware of their existence and type when checking the `hermes-eslint` package. Note that we also tweak some of the logic around Hermes parser options to match the new Flow types.

We should configure Flow to ignore the `scripts` and `dist` directories, as these contain JS files that are not source code that should be checked.

Finally, we must work around an issue with Flow and yarn workspaces. Yarn workspace will hoist packages in the workspace to the top level `node_modules`, causing Flow to produce module resolution errors when these packages (e.g. `hermes-parser`) are required. We can work around this by adding a `module.name_mapper` entry to the flowconfig that resolves the `hermes-parser` module to the type definitions for its exports.

Reviewed By: pieterv

Differential Revision: D26990854

fbshipit-source-id: 1c6d10a8ae5e118d99e81bb6d0a3968163983005