Add Flow types to scope manager Scopes
Summary:
This diff adds Flow types to the `scope.js` file in `hermes-eslint`'s scope manager.

For the most part this simply involves adding type annotations, however there are a few locations where values are nullable where we need to either add optional chaining, or slightly refactor to preserve a non-null type refinement.

We also add a new `__isClosed` property instead of setting `__referencesLeftToResolve` (which is usually an Array) to null to allow for proper Flow typing.

Reviewed By: pieterv

Differential Revision: D26999137

fbshipit-source-id: 3f306e43caf636b4c57b18263258c375ab583487