Avoid passing GCHermesValue& around in JSObject
Summary:
Instead of passing around references to GCHermesValue, accept and
return HermesValues from get/set functions in JSObject. This reduces
the possibility of accidentally holding dangling references into a
relocated JSObject, which is going to be more of a risk with HV32
(since many more places will be able to allocate).

This change also lets us remove references from the API for
ArrayStorage.

Reviewed By: avp

Differential Revision: D26998350

fbshipit-source-id: defa243c3784c8013eaa3109e84eb5219b06c981