Add documentation on heap snapshot support
Summary:
Chrome heap snapshots are now the best way to understand Hermes's memory usage.
We want to make sure everyone knows how to use these well to identify areas
of improvement for their app.

Add some documentation including images and videos showing how to do various
basic tasks in heap snapshots. Include some Hermes-specific information (for example
about Environments) to help people better understand how Hermes works.

Videos are stored at `assets.hermesengine.dev` to avoid increasing the size of the repository
when clones are made. A side bonus is that this means the videos work well even when viewed
from Github or with Markdown preview. It might be hard to version the videos, but
it doesn't seem likely that they'll be changed often (and we can always just add `v2` as a suffix
to the video file name).

Images are stored in-repo for simplicity and they are much smaller than videos to download.

Reviewed By: Huxpro

Differential Revision: D26180717

fbshipit-source-id: 5e78a10772be1936ceedfd7bbf10fd8ef8b84895