Define clang test coverage variables in CMake
Summary:
It would be nice to get some test coverage data for our source code in Hermes.
To do so, take advantage of some clang flags: `-fprofile-instr-generate -fcoverage-mapping`.
These two flags make a binary compiled with them output profiling data to
a `profraw` file, which can be merged and then displayed as HTML or textual formats.

Reviewed By: neildhar

Differential Revision: D27055012

fbshipit-source-id: fd00ec80914460ac1fb97b60c6c85801217df758