Remove any usage of FASTCOMP by default in builds (#472)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/472

Since fastcomp was deprecated in emscripten v2 and above,
change the default in most places (configure.py and CMake) to be
"upstream".
This shouldn't have any effect on existing builds because they all
manually specified upstream anyways.
This change will prevent accidentally using fastcomp in the future,
although it's still possible to use it if you downgrade to emscripten v1.*

Reviewed By: neildhar

Differential Revision: D27128367

fbshipit-source-id: 4ba5f0f2071961c2217491e18c36a2e48712e1d2