Add LTO to Release and MinSizeRel builds in CMake (#473)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/473

When building with the CMake build mode "Release" or "MinSizeRel", turn
on LTO in compiler flags and linker flags.
This will impact builds of binaries like `hermes` and `hermesc`, as well as the
libraries packaged in our OSS builds.

There are a couple advantages to turning on LTO, one of which is a size difference:
The hermes binary changes from 3.4 MB to 2.5 MB for -Os (MinSizeRel)
and from 3.9 MB to 3.3 MB with -O3 (Release)
The second advantage is performance, although in this case Octane scores didn't
seem to be statistically significantly affected.

These benefits can extend to the builds for apps like `libhermes.so`, but the benefits
aren't as noticeable.
If someone statically linked their app libraries they might get more benefit out of it.

Differential Revision: D22529505

fbshipit-source-id: 88af10e7f4df7b44b0e631882856424ab4ad537e