Refactor JSObject computed descriptor API.
Summary:
The computed descriptor API needs to handle Proxy and HostObject
properly in order to be correct.

Split the existing JSObject functions which take `ComputedPropertyDescriptor`
as input into standard and `Internal` variants. The `Internal` variants
must only be called when the caller is certain that the target is not
a Proxy or HostObject.

The lookup of a `ComputedPropertyDescriptor` may generate a new
`SymbolID`. This `SymbolID` must then be used again when retrieving
the actual value, and must not be freed between the two calls.
To facilitate this, add a `tmpSymbolStorage` handle parameter to both
functions which is used as temporary storage. The same temporary storage
being passed to both `getComputedDescriptor` and `getComputedSlotValue`
ensures that the `SymbolID` will be valid.

The `SymbolID` itself is stored and read from the descriptor directly
by the functions.

Reviewed By: tmikov

Differential Revision: D25138513

fbshipit-source-id: 7846fd4fbc65a2f1b16308a2a83d87bf344805f9