Use the CircleCI Android orb (#464)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/464

CircleCI defines a convenience "orb" that has a lot of stuff built in:
* A system image with Android SDK and NDK pre-installed
* Common steps for creating an AVD, starting it, and waiting for it
* Comes with node.js preinstalled

The preinstalls means there's no lengthy download step, that often
fails flakily.
This orb also has commands for persisting a build and gradle cache,
which I haven't added yet. It might be good to add those in the future
to speed up the jobs.

I think some other things in this script can be simplified as well, such
as downloading and pinning cmake and ninja to specific versions, but
I don't want to mess with them yet.

Full documentation of what the Orb can do is here:
https://circleci.com/developer/orbs/orb/circleci/android

Differential Revision: D25786818

fbshipit-source-id: 7f7f0005e8abcd0df14aacefb26ece864bf7e577