Use the CircleCI Android orb for test-e2e (#478)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/478

Change the `test-e2e` job from running on a physical mac machine
to using a Linux docker container via the Android orb.
This has a few convenience functions, including:
* Pre-installed node.js
* Easy installation of ndk
* Creating AVDs and starting the emulator

Note that unfortunately due to KVM and hardware acceleration not being available
on CircleCI docker machines, ARMv7 needs to be used instead of x86, which
can slow down the test. Because of this, the API had to be downgraded to v25, which
was the most recent version with an armv7 system image available.

Differential Revision: D27302251

fbshipit-source-id: 9dd42875c0b0f1132b852e5d5cc449eec4b8287e