Use get/setDirectSlotValue when index is statically known
Summary:
There are many places where internal properties used by an object are
known to be in direct property slots at compile time. In those cases,
we should prefer using get/setDirectSlotValue since it avoids a
dynamic check for whether the property is in a direct slot. It also
lets us remove the runtime argument from certain places.

This is also useful for helping keep track of changes from HV32,
since the current get/setInternalProperty mechanism is harder to keep
track of, because it does not perform any validation.

Reviewed By: avp

Differential Revision: D27309770

fbshipit-source-id: a74ec1d81f8ed3c694ef88f93a31fb62f46a4706