Reduce handle allocations during Array creation
Summary:
`Arguments::create` no longer needs to allocate a Handle for the
indexed storage, because the indexed storage can be allocated and
assigned after the Arguments object has been created.

For `JSArray::create`, we can create a `Handle<JSArray>` instead of a
`MutableHandle<StorageType>` by reversing the allocation order. This
means that we can return a `Handle<JSArray>` from `JSArray::create`,
which saves callers having to allocate a Handle themselves.

Reviewed By: dulinriley

Differential Revision: D27355590

fbshipit-source-id: c5cc4657368c4700f35972ce02dfb4c6367c8b9f