Implement dummy SmallHermesValue
Summary:
Implement a version of SmallHermesValue that provides the API of a
32-bit HermesValue but internally is just a 64-bit HermesValue.

Once a 32 bit version is implemented, we can use a typedef to select
between the two based on build flags. The functions themselves are
kept in HermesValue itself, because this lets us write templates that
operate on both types (e.g. GCHermesValueBase).

Reviewed By: dulinriley

Differential Revision: D25985115

fbshipit-source-id: e17c5c0329c2c7353922ee93b1f407d2d1a36972