Use DiagnosticStat for tracking overall stats
Summary:
Use the DiagnosticStat type introduced earlier to replace all stat
tracking in the Heap Size Diagnostic acceptor.

The key differences in output are that empty entries are now
excluded, and that the order of things has changed. Apart from
simpler code, the new version is also less error prone, since we
don't need to maintain several custom formulas to compute certain
stats.

In implementing this, I also discovered that our current computed
root size is incorrect on 64 bit platforms, because root pointers are
assumed to be the same size as GCPointers.

Reviewed By: dulinriley

Differential Revision: D27458504

fbshipit-source-id: 6bed56bf85e77648c1f657f656f6dfc5a7d76a44