Implement 32 bit HV representation
Summary:
# Implement HermesValue32

Implement a compressed 32-bit representation of HermesValues. This
supports storing everything a full sized HermesValue can support
except for native values.

## Properties
The `HermesValue32` type is trivially copyable and is just as cheap
to construct, copy, or destroy as a full sized HermesValue. It pegged
to the size of a `GCPointer`, which means that it is 32-bits on
32-bit systems and 64-bit systems with compressed pointers enabled.

The type of value is stored in a 3 bit tag. The size of the tag is
limited by the heap alignment, because we need to be able to
accommodate an entire pointer inline in the class.

## Pointers
Pointers are stored using the raw internal representation used in
GCPointer. A pointer may be a `String`, `Object`, or `BoxedDouble`.
All pointers must point to places on the heap, since they used
compressed pointer encoding. Reading or writing a pointer requires a
PointerBase.

## Numbers
Unlike a 64-bit `HermesValue`, `HermesValue32` cannot store a full
double inline. Instead, we store only doubles that can be represented
as small integers inline. In this diff, that means that only values
that can be represented as signed integers of size `kNumSmiBits` can
be stored inline.

Values that cannot be stored inline have to be heap allocated. They
are allocated as `BoxedDouble` objects in the heap. In order to
maintain trivial copyability, the double is immutable, so copying a
pointer to a `BoxedDouble` is always safe. This means that every
double encoding requires having a `GC*` and can cause an allocation.

## GC Safety
For convenience, a `encodeHermesValue` function exists, which makes
it easy to convert a `HermesValue` to a `HermesValue32`. If the value
being encoded is a non-SMI double, then it will have to be allocated
on the heap. The API of this function guarantees that encoding
pointer values will not allocate, because an allocation could
invalidate the pointer value we are trying to encode.

To enforce this requirement, we allocate all numbers on the heap when
Handle-SAN is disabled. In addition, any non-pointer value given to
`encodeHermesValue` will also cause an allocation, to maximise
coverage of codepaths that store `HermesValue`s.

There are only two functions that can allocate: `encodeHermesValue`
and `encodeNumberValue`.

## Converting Back
There are two functions to convert a `HermesValue32` to a
`HermesValue`: `unboxToHV` and `toHV`. Both functions require a
`PointerBase` to decompress the pointer, but are otherwise fast and
safe.

`unboxToHV` is the only one meant to be used in practice, since it
will also decode boxed doubles back to HVs. `toHV` is meant more for
debugging and diagnostics, and it simply converts boxed doubles to
their corresponding object pointers. This is useful for code reuse in
the GC.

Reviewed By: dulinriley

Differential Revision: D26447669

fbshipit-source-id: 6cefa777a431a36610838765ec6d13f28e245baa