Make Hades the default for CMake builds (#489)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/489

Now that Hades has been working well for a while and hasn't had any major bugs,
it's time to introduce it to the world.
Setting this as the default GC for CMake builds. This means after this change
is merged, any OSS build of Hermes will be using Hades.
When Hermes makes its next release cut (v0.8) this will be included.

Reviewed By: neildhar

Differential Revision: D27606800

fbshipit-source-id: 16ebd30c1b876116b643d2835b6720ea78688c3d