Add documentation for GenGC
Summary:
GenGC is quite old now, but we had no high level design documentation
of how it works.
This documentation explains most of the key pieces of the GC, many of which
apply to any Hermes GC:
* Heap Segments
* Metadata for finding pointers
* Generations (young and old)
* Write barriers and card tables
* Mark, Sweep, Compact phases
* Mark bits
* Weak references and WeakMap

Later I'll add Hades documentation, and refer to this document to establish some
baseline knowledge so Hades can focus on the differences.

The intended audience are people who are unfamiliar with general garbage collection
principles, so there's some basic explanation of the goals of a garbage collector.
There are a couple references to classes and functions in the GC to make it easier
to find the relevant source code.

There aren't any fun pictures, videos, or diagrams yet, mostly because writing text
is a lot easier. Perhaps later we can add visual stuff if someone wants it.

Reviewed By: neildhar, avp

Differential Revision: D27249383

fbshipit-source-id: 83866fa07e7b03fc96de8d350a6f847a13b11ab2