Remove read barrier from SamplingProfiler path
Summary:
The SamplingProfiler was executing read barriers on weak refs, but it might
have interrupted the mutator in the middle of its own write barrier.
To avoid a race condition with the marking queue used by the barriers, disable
read barriers in the sampling profiler.
To avoid the case where the Domain is garbage collected, add a new GCCallback
function `markRootsForCompleteMarking`, which is optionally called by a GC.
This is used only by Hades, as GenGC doesn't have a weak ref read barrier.

Runtime implements this by remarking the domains collected by the sampling profiler.
Note that this avoids many of the signal handler issues like not being allowed to allocate
memory.

Reviewed By: neildhar

Differential Revision: D27601473

fbshipit-source-id: cc1467933d5e08bb64455b5fafec475e162da2df