Improve build times using unity builds
Summary:
Allow us to use CMake Unity builds, which batch cpp files into groups
of 8 and then compile them together. This means we don't have to
compile shared header files as many times, and saves on linking
time.

Exclude LLVH so we don't have to make changes to it, and because it
is unlikely to contribute to incremental build times anyway. Exclude
some of our unit tests because they have conflicts with free
functions.

If we now want to use unity builds, we can just add
`-DCMAKE_UNITY_BUILD=ON` when configuring.

Reviewed By: dulinriley

Differential Revision: D27555688

fbshipit-source-id: dfb9092298472f71201a8598935b1c51a5ef315f