Optimise unboxToHV
Summary:
`unboxToHV` turns out to be a pretty significant part of execution
now. We can optimise it by using a very similar strategy to the one
used in `HermesValue`. Essentially, it works by using one extra bit
for the tag. However, for types that cannot afford to lose the extra
bit to the tag, we instead map them to two 4 bit tags (representing
both possible values of the extra bit).

This lets us execute `unboxToHV` as a single switch over all the
extended tag types, instead of needing a separate switch for the
previous `ValueTag` mechanism.

Reviewed By: dulinriley

Differential Revision: D27643338

fbshipit-source-id: 1825302092651475ab5e6103b9de7b6016fe43c1