Move inline GCBase functions to GCBase.cpp
Summary:
These functions currently have to be recompiled for a huge number of
translation units in the VM, since they get included anywhere the GC
or Runtime are used.

Since it isn't perf critical to inline them (they are mostly for
instrumentation), we can move them all into GCBase.cpp.

Reviewed By: dulinriley

Differential Revision: D27814568

fbshipit-source-id: 1ccf9d9fdef75b408f6b15d7c267e6d43b288533