Populate SegmentID based off SourceURL for classic bundles
Summary:
Given the build time regression preventing Hermes Modules to ship in debug builds in the near future (more info: https://fb.workplace.com/groups/684200151742597/permalink/1876367169192550), we need to break our dependency on this work for the Jest E2E coverage collection. Because the SegmentID isn't populated in classic builds, the API will return bogus data for us unless we're using Hermes Modules.

To circumvent this, I've created a workaround in this diff that attempts to extract the SegmentID from the SourceURL. This is less than ideal obviously, but without it we cannot ship coverage collection for FB4a, which means we can't do landblocking test runs on React Native product changes. As the usage of Jest E2E is growing, this is causing increasingly large oncall loads for both us (Failing tests blocking OTA updates) and product teams (Oncall having to investigate, bisect and fix failures in the test). For that reason, I think it's worth putting this workaround in place until we have Hermes Modules rolled out successfully.

Reviewed By: makovkastar

Differential Revision: D27293987

fbshipit-source-id: 0090e30dc956ecfd375f496904351f48a6ca1a1b