Support lookup from snapshot ID to object
Summary:
In order to support querying objects in a heap snapshot for things
like object properties, we need to support mapping from an ID back to
an object.

The particular interaction that causes this is hovering over an object in
the snapshot for a second while the debugger is attached.

Maintaining a fast lookup table would be very wasteful of memory for the
< 100 objects that someone would query. So instead, do a lazy search
of all objects whenever one is requested. In practice when I tried this
on a small heap (< 1 MB) it didn't take long at all. I'd rather have the UI
stall then use up 2x the memory for tracking objects.

After the first time this occurs, the ID -> object mapping is cached and kept
up to date until the object dies. That means going back and forth between
a few objects will be fast.

If people in the wild complain about the performance of the UI, we might need
to revisit this decision, but I doubt it'll be a problem in practice.

Reviewed By: neildhar

Differential Revision: D27834673

fbshipit-source-id: 844b1ec13c80176b9fdb0c44616273ecf23ec08a