Handle non-zero VM allocations
Summary:
Hades currently implicitly relies on memory being zeroed out when it
is allocated from the OS, since it doesn't manually initialise the
card table and mark bits. When the segment is not zeroed out, this
puts the heap in an invalid state and causes asserts to fire.

GenGC's segment initialisation code already does this, so move that
initialisation code into the base class shared by both Hades and
GenGC.

Reviewed By: tmikov

Differential Revision: D27889252

fbshipit-source-id: 97f0db1434f6b2c83a1288ca36767f95bd8d1d76