Enable trimming in concurrent mode for blocking collections
Summary:
Trimming is currently disabled during sweeping in concurrent Hades,
because it is unsafe to concurrently trim cells. However, there are
scenarios where we are actually running the sweeper on the mutator,
such as when a user triggers a GC or when we get a memory warning.

In those cases, we can actually trim arrays, which can be
particularly valuable under memory pressure.

Reviewed By: dulinriley

Differential Revision: D27932791

fbshipit-source-id: a77a266bb1e5bc84e8ec6634a46e93a3fd28f670