Use union to store forwarding pointer in GCCell
Summary:
The `VTable` pointer in the `GCCell` can currently be used to store
either an actual `VTable` pointer, or a forwarding pointer (during a
collection). Always treating this as a `VTable` pointer ends up
requiring repeatedly casting the value and using `const_cast` (since
the `VTable` is `const` but the `GCCell` is not). Rewriting this as a
union makes the code a little simpler.

This is necessary so that we can potentially replace the `VTable`
pointer with a `CellKind` and size.

Reviewed By: dulinriley

Differential Revision: D27990735

fbshipit-source-id: 676023ac4768cf8035189a7e538376e153200b4f