Always fill in SegmentedArray::increaseSize
Summary:
There is currently only one case where we do not fill the array after
increasing its size, which occurs in `growLeft` when we are not on a
concurrent GC. There are a few reasons to remove this special case
altogether:
1. `growLeft` is relatively rare, and has other performance
bottlenecks.
2. The code is tricky, for instance, we currently seem to fill the
start of the new array in `growLeft` twice for concurrent Hades.
3. We keep the invariant that anything within the size (that is
therefore reachable by a concurrent GC), is always initialised.
4. We can change the `uninitialized_copy` in `growLeft` to just be a
`copy`. This is technically more expensive, but it creates the
invariant that `uninitialized_copy` is only ever called on memory
that is not reachable by a concurrent GC. This lets us enable the
pointer overloads of `uninitialized_copy` for Hades in the next diff.

Reviewed By: dulinriley

Differential Revision: D28209771

fbshipit-source-id: c911dff5ee5a12d798529a56e1f8bc77eb5d02c5