Use open ranges in card table
Summary:
Closed ranges make the code more prone to off-by-1 errors, since
there are multiple places where we need to add/subtract 1 to get the
right value.

This is particularly true for pointer ranges (like in
`dirtyCardsForAddressRange`). The current behaviour is
counterintuitive since if `high` represents a card boundary, we will
end up dirtying a card that does not actually fall inside the range.

This also currently complicates `updateAfterCompaction`, where we
need to rely on the placement of the card table at the start of a
segment to avoid underflowing.

With this change, the tricky behaviour is constrained to managing the
right calls to `addressToIndex`. In future, we could even split that
into two different functions, for the start and end of a range.

Reviewed By: dulinriley

Differential Revision: D28232756

fbshipit-source-id: 0d2a19d4fc47ef3b0e6bcea8d6ad2715776a4e6f