Allow efficient uninitialized_copy in Hades
Summary:
Previously, it was not possible to enable the pointer overload of
`uninitialized_copy` on Hades since we don't really get atomicity
guarantees from `memcpy`. However, now that we have an invariant that
the memory passed to `uninitialized_copy` must not be reachable by
the GC, we don't need the atomicity.

Reviewed By: dulinriley

Differential Revision: D28210620

fbshipit-source-id: 86a7b329f358ee6d065bc6c9296a46d5a68f2db1