Stop trimming during Hades compaction
Summary:
Simplify compaction by removing cell trimming from it. If necessary,
we can still trigger trimming by doing a blocking collection on the
mutator thread.

The main purpose of this is to reduce the size of `CopyListCell`.
Later in the stack, I'll remove the overhead of including sizes in
variable sized cells, so this will ensure that `CopyListCell` is
still as small as the smallest possible cell.

The other option I evaluated is to start enforcing minimum element
counts for cells like `ArrayStorage` so that they are always at least
as large as the `CopyListCell`. That seems to perform worse and
consume more memory based on benchmarks.

Reviewed By: dulinriley

Differential Revision: D28316035

fbshipit-source-id: 0a4b55a5e780eb064e7ba91c4aadb597a0f6fb53