Skip last segment for compaction
Summary:
Exclude the last segment from compaction. This ensures that we do not
try to compact in very small heaps that only have a single segment
(since that would remove the only segment from the heap).

It also ensures that the compaction heuristic doesn't always select
the most recently added segment, which is always going to be the
smallest segment, but is not necessarily the best compaction
candidate (since it is unlikely to be fragmented).

This also avoids breaking the trimming tests in the next diff, since they
have heaps of just one segment, that then gets selected for
compaction. Without trimming in compaction, the array cells end up
not getting trimmed.

Reviewed By: dulinriley

Differential Revision: D27249199

fbshipit-source-id: 19dd081011d7cef314221ccffcdc7cc188d9234c