Split WeakRef and WeakRoot marking
Summary:
We have an optimisation in our cell marking code that avoids checking
whether a cell has a `WeakRef` marking function if the acceptor being
used does not derive from `WeakRefAcceptor`. In order to mark and
update `WeakRoot`s (prior to this stack, it was just for compaction),
the Hades YG `EvacAcceptor` derives from `WeakRootAcceptor`. This
ends up incurring the cost of checking the `VTable` for every YG cell
we mark.

In reality though, we only ever mark weak roots with the acceptor, so
the `WeakRef` checks are not necessary. Since there are no more
`WeakRef`s in our weak roots, we can split up the acceptors, and
remove the unnecessary check from YG collections.

Reviewed By: dulinriley

Differential Revision: D28339574

fbshipit-source-id: c08ca0d1e9faf8cf142ff9add3fae61625f11307