Fix warnings about implicitly defined copy and assignment
Summary:
Various classes defined an `operator=` without a copy ctor, or vice versa.
This causes a deprecation warning in new versions of the compiler,
so add a corresponding definition for each.

Delete the `GCPointer` copy ctor, which requires using
`CompressedPointer` instead in a few places and forces an extra
`vmcast` because `CompressedPointer` doesn't retain GCCell type info.

There is also one patch to LLVM.

Reviewed By: neildhar

Differential Revision: D28551199

fbshipit-source-id: b306fb71ad6dbb7d8cd27ae0cc72f982f686a462