Split out bytecode initialisation of JSRegExp
Summary:
Split out the version of `JSRegExp::initialize` that takes existing
bytecode into a separate function. This version cannot throw an
exception because the bytecode and pattern are assumed to be
correctly formed.

This involves deleting the current check that the bytecode size fits
in a `uint32_t`, which wasn't particularly meaningful anyway because
the regex compiler uses 32 bit offsets in multiple places.

Reviewed By: avp

Differential Revision: D28583013

fbshipit-source-id: 93d216aebcf2107261f8df98ced409f34692023d