Detect host platform instead of target for InternalBytecode Cmake
Summary:
There is a switch for choosing the right concatinator for concating js file together
in `lib/InternalBytecode/CMakeLists.txt`. When crosscompiling (eg compiling Android on Windows)
it chooses the target platform (not WIN32) despite being launched on Windows.
CMAKE_HOST_WIN32 var is suggested to be used instead to make the proper detection.