Remove capacity from ArrayStorage
Summary:
This diff aims to solve two problems.

First, HV32 and compressed cell headers mean that using even-numbered
values of `capacity_` can be inefficient. This is because the size of
an empty array storage is only 12 bytes, so when we add an even
number of 4 byte properties, one slot gets wasted due to heap
alignment.

Second, the `capacity_` slot for a cell is somewhat duplicative,
since the capacity can be calculated as a function of the overall
size of the cell (which is already recorded).

To address both of these, remove the `capacity_` member and instead
always use the allocated size of the cell to determine how many slots
can fit inside it. This saves memory and removes the possibility of
wasted slots.

Reviewed By: kodafb

Differential Revision: D28669251

fbshipit-source-id: 21edaa0a01be4d034e8ddfecb8c01182daac9476